refactor methods, add stat_tracker_helper for methods across modules
