+++ lib/stat_tracker.rb
@@ -9,6 +9,7 @@ require_relative '../modules/team_statables'
 require_relative '../modules/team_stat_helper'
 require_relative '../modules/season_stat_helper'
 require_relative '../modules/game_stat_helper'
+require_relative '../modules/stat_tracker_helper'
 require_relative '../modules/season_statables'
 
 class StatTracker
@@ -19,6 +20,7 @@ class StatTracker
   include TeamStatHelper
   include SeasonStatHelper
   include GameStatHelper
+  include StatTrackerHelper
   attr_reader :games, :teams, :game_teams
 
   def initialize(games = {}, teams = {}, game_teams = {})
