+++ modules/season_stat_helper.rb
@@ -1,17 +1,5 @@
 module SeasonStatHelper
 
-  def all_season_hash(team_id)
-    all_games_won_by_season(team_id).merge(all_games_played_by_season(team_id)) {|season, won, played| won / played.to_f}
-  end
-
-  def all_seasons_ary
-    all_seasons_ary = []
-    @games.values.map do |game|
-      all_seasons_ary << game.season
-    end
-    all_seasons_ary.uniq
-  end
-
   def all_games_played_by_season(team_id)
     all_season_games = Hash.new
     all_seasons_ary.each do |season|
@@ -35,9 +23,39 @@ module SeasonStatHelper
     all_season_won_games
   end
 
+  def games_per_season_type(team_id, post_reg, all_games = false)
+    season_games = Hash.new(0)
+    all_seasons_ary.each do |season|
+      season_games[season] = @games.values.count {|g| g.season == season && g.type == post_reg && (g.home_team_id == team_id || g.away_team_id == team_id)}
+    end
+      season_games.each do |season, games|
+        season_games.delete(season) if games == 0
+      end
+  season_games
+  end
+
+  def games_won_per_season_type (team_id, post_reg)
+    season_games = Hash.new(0)
+    all_seasons_ary.each do |season|
+      season_games[season] = 0
+      @games.values.each do |g|
+        g.season == season && g.type == post_reg &&
+        (g.home_team_id == team_id || g.away_team_id == team_id) &&
+        if g.home_team_id == team_id &&
+          home_team_won?(g)
+          season_games[season] += 1
+        elsif g.away_team_id == team_id &&
+          away_team_won?(g)
+          season_games[season] += 1
+        end
+      end
+    end
+    season_games
+  end
+
   def season_win_percentages_type(team_id, post_reg)
     seazy = games_won_per_season_type(team_id, post_reg).merge(games_per_season_type(team_id, post_reg)) {|season, won, played| won / played.to_f}
     seazy.transform_values{|v| v.round(2)}
   end
-  
+
 end
